+++ CHANGELOG.md
@@ -57,6 +57,7 @@
 - Added support for `Series.dt.is_year_start` and `Series.dt.is_year_end`.
 - Added support for `Series.dt.is_quarter_start` and `Series.dt.is_quarter_end`.
 - Added support for `Series.argmax` and `Series.argmin`.
+- Added support for `Index.argmax` and `Index.argmin`.
 
 #### Improvements
 - Removed the public preview warning message upon importing Snowpark pandas.